LPS-67836 Hugo SF - force module names to be consistent. Oherwise, I could have an app named "bob" and the module name could be "jane" instead of "bob-xyz". This is not like foundation, but more like foundation/mobile-device-rules.
