+++ modules/apps/social-office-upgrade/social-office-upgrade-association/build.gradle
