+++ modules/apps/social-office-upgrade/social-office-upgrade-association/src/main/java/com/liferay/social/office/upgrade/association/internal/constants/RoleConstants.java
@@ -12,20 +12,15 @@
  * details.
  */
 
-package com.liferay.social.office.association.internal.constants;
+package com.liferay.social.office.upgrade.association.internal.constants;
 
 /**
+ * @author Jonathan Lee
  * @author Eudaldo Alonso
  */
-public class SocialOfficeConstants {
+public class RoleConstants
+	extends com.liferay.portal.kernel.model.RoleConstants {
 
-	public static final String LAYOUT_SET_PROTOTYPE_KEY =
-		"layoutSetPrototypeKey";
-
-	public static final String LAYOUT_SET_PROTOTYPE_KEY_USER_PRIVATE =
-		"layoutSetPrototypeKey_UserPrivate";
-
-	public static final String LAYOUT_SET_PROTOTYPE_KEY_USER_PUBLIC =
-		"layoutSetPrototypeKey_UserPublic";
+	public static final String SOCIAL_OFFICE_USER = "Social Office User";
 
 }
