+++ modules/apps/social-office-upgrade/social-office-upgrade-association/src/main/java/com/liferay/social/office/upgrade/association/internal/model/listener/GroupModelListener.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.social.office.association.internal.model.listener;
+package com.liferay.social.office.upgrade.association.internal.model.listener;
 
 import com.liferay.expando.kernel.model.ExpandoTableConstants;
 import com.liferay.expando.kernel.service.ExpandoValueLocalService;
